requirements.txt
Flask==2.2.2
SQLAlchemy==2.0.0
requests==2.28.1
python-dotenv==0.21.0
openai==0.27.0
minio==7.1.1
